chore: Code reorganization
