+++ src/lexer/lexer.ts
@@ -2,7 +2,6 @@
 
 import * as moo from "moo";
 import type { Conditional } from "src/types";
-import { Parser } from "./parser";
 
 export const TAG_REGEX =
     /(?:\d+[Dd])?#(?:[\p{Letter}\p{Emoji_Presentation}\w/-]+)(?:\|(?:[+-]))?(?:\|(?:[^+-]+))?/u;
@@ -21,6 +20,86 @@ export const OMITTED_REGEX =
 export const CONDITIONAL_REGEX =
     /(?:=|=!|<|>|<=|>=|=<|=>|-=|=-)(?:\d+(?:[Dd](?:%|F|-?\d+|\[\d+(?:[ \t]*[,-][ \t]*\d+)+\]|\b))?)/u;
 
+type ParseContext = {
+    associativity: "left" | "right";
+    precedence: number;
+};
+
+class Parser {
+    table: Record<string, ParseContext>;
+    constructor(table: Record<string, ParseContext>) {
+        this.table = table;
+    }
+    parse(input: LexicalToken[]) {
+        const length = input.length,
+            table = this.table,
+            output = [],
+            stack = [];
+        let index = 0;
+
+        while (index < length) {
+            let token = input[index++];
+
+            switch (token.value) {
+                case "(":
+                    stack.unshift(token);
+                    break;
+                case ")":
+                    if (
+                        input[index] &&
+                        input[index].type == "dice" &&
+                        /^d/.test(input[index].value)
+                    ) {
+                        input[index].parenedDice = true;
+                    }
+                    while (stack.length) {
+                        token = stack.shift();
+                        if (token.value === "(") break;
+                        else {
+                            output.push(token);
+                        }
+                    }
+
+                    if (token.value !== "(")
+                        throw new Error("Mismatched parentheses.");
+                    break;
+                default:
+                    if (table.hasOwnProperty(token.value)) {
+                        while (stack.length) {
+                            const punctuator = stack[0];
+
+                            if (punctuator.value === "(") break;
+
+                            const operator = table[token.value],
+                                precedence = operator.precedence,
+                                antecedence =
+                                    table[punctuator.value].precedence;
+
+                            if (
+                                precedence > antecedence ||
+                                (precedence === antecedence &&
+                                    operator.associativity === "right")
+                            )
+                                break;
+                            else output.push(stack.shift());
+                        }
+
+                        stack.unshift(token);
+                    } else {
+                        output.push(token);
+                    }
+            }
+        }
+
+        while (stack.length) {
+            const token = stack.shift();
+            if (token.value !== "(") output.push(token);
+            else throw new Error("Mismatched parentheses.");
+        }
+
+        return output;
+    }
+}
 export interface LexicalToken extends Partial<moo.Token> {
     conditions?: Conditional[];
     parenedDice?: boolean;
@@ -29,6 +108,30 @@ export interface LexicalToken extends Partial<moo.Token> {
 }
 
 class LexerClass {
+    constructor() {
+        this.parser = new Parser({
+            "+": {
+                precedence: 1,
+                associativity: "left"
+            },
+            "-": {
+                precedence: 1,
+                associativity: "left"
+            },
+            "*": {
+                precedence: 2,
+                associativity: "left"
+            },
+            "/": {
+                precedence: 2,
+                associativity: "left"
+            },
+            "^": {
+                precedence: 3,
+                associativity: "right"
+            }
+        });
+    }
     lexer = moo.compile({
         WS: [{ match: /[ \t]+/u }, { match: /[{}]+/u }],
         table: TABLE_REGEX,
@@ -109,30 +212,6 @@ class LexerClass {
     }
     public setDefaultFace(face: number) {
         this.defaultFace = face;
-    }
-    constructor() {
-        const exponent = {
-            precedence: 3,
-            associativity: "right"
-        };
-
-        const factor = {
-            precedence: 2,
-            associativity: "left"
-        };
-
-        const term = {
-            precedence: 1,
-            associativity: "left"
-        };
-
-        this.parser = new Parser({
-            "+": term,
-            "-": term,
-            "*": factor,
-            "/": factor,
-            "^": exponent
-        });
     }
     parse(input: string): LexicalToken[] {
         const tokens = Array.from(this.lexer.reset(input));
